+++ include/async_mqtt5/impl/unsubscribe_op.hpp
@@ -75,9 +75,10 @@ public:
 			topics, props
 		);
 
-		auto max_packet_size = _svc_ptr->connack_prop(
-			prop::maximum_packet_size
-		).value_or(default_max_send_size);
+		auto max_packet_size = static_cast<size_t>(
+			_svc_ptr->connack_prop(prop::maximum_packet_size)
+				.value_or(default_max_send_size)
+		);
 		if (unsubscribe.size() > max_packet_size)
 			return complete_post(
 				client::error::packet_too_large, packet_id, topics.size()
